start of month exception
some months began with a 0 first day. so changed use effect to trigger on day 2, and set end of month = 0 to end of month = 1 (which governs weeks running over the end of the month) this way, the fix doens't trigger a new month with the use effect. checked for a year and it was ok. the calandar breaks down a year or so in