feature lesson attachment (#13)
* Moved to Component, Added state

moved lesson attachment to a new component file lesson-create-attachment.js. also added state and event handler for selecting files. e.target.files creates a button that opens a file navigator/selector

* Expanded min-width

lessonControlAttachment min-width =295px to show file name

* Attachments list, one behind

Attachments are added to Lesson Attachment box but are one behind

* attachArray adds <p>

Using a toggled state for isSelected in the handleAddAttachment, the attachArray now displays each attachment as a <p> after the button is clicked

* remove attachment in progress

using handleRemoveAttachment with slice and identifying the position of the attachment of attachArray with id=i (from the map) will remove an element. still in the works

* lesson attachment complete

lesson attachment feature now adds lessons to the list, creating an array attachArray. should also delete an item once hooked up to the back end

* feature complete

Attachment list as array is passed to lesson-create-control through lessonDataSlice. Lesson is created on click