feature edit edit back (#23)
* Edit Handler, Fields to Input

Added handler which switches state of isEditOn, when clicked form fields are changed from <p> to <input type=;text'>

* Lesson Change in Edit Control

added useEffect to update states of lesson data fields when user selects a different lesson

* Selected lesson edit to lessonDataSlice

handleEditLesson passes a new object with the index and updated lesson data to lessonDataSlice

* End of day

* Edit updates table

In editLessonData reducer made the lesson date into an object to match the objects in lessonDataArr. The splice now updates the lesson table

* Edit details on Save

Added setIsEditOn(!isEditOn) to handleEditLessonData so the lesson details goes back to the format without input fields. the lesson data in that is up to date

* Lesson Attachment in Edit

added <LessonCreateAttachment /> to lesson-edit-control to enable editing attachments. It renders on load and clears the values of the attachments. Investigating why and how I set that up. maybe need to set that useEffect as a function and run the function in the handleClearAttachment?

* end of day

* lesson Attachment edit update

Created setAttachmentList in lessonControlSlice to pass the attachment list after edit to the Lesson Datails and Lesson Table

* lesson edit control update

passed lessonAttachmentList instead of singleLessonData[5] into LessonEditAttachment from lesson-edit-controls. Now if the user clicks Edit Lesson again it will show the updated lesson attachments (not the original ).

* Removed commented out code, console.logs, unused imports

* lesson-edit-control attachment update

changed the lesson attachment map from the useSelector value (lessonAttachmentList_ to the local state lessonAttachment to allow for the first render (as lessonAttachmentList is undefined). setLessonAttachment(lessonAttachmentList) in a useEffect which has lessonAttachmentList as dependency array

* Styling Edit Lesson